Refactor Resource wrapper class
